requirements.txt
openai
python-dotenv
google-generativeai
requests
flask
